+++ glfw/cocoa_window.m
@@ -1927,7 +1927,12 @@ void _glfwPlatformDestroyWindow(_GLFWwindow* window)
     window->ns.view = nil;
 
     [window->ns.object removeGLFWWindow];
-    [window->ns.object close];
+    // Workaround for macOS Tahoe where if the frame is not set to zero size
+    // even after NSWindow::close the window remains on screen as an invisible
+    // rectangle that intercepts mouse events and takes up space in mission
+    // control. Sigh.
+    [window->ns.object setFrame:NSMakeRect(0, 0, 0, 0) display:NO];
+    [window->ns.object close]; // sends a release to the NSWindow so we dont release it here
     window->ns.object = nil;
 }